MacOS Support (#615)
* working macOS build

* build app bundle

* fix fonts

* fix fonts when copied to /Applications

* Made more modification on newer files to ensure macos compatibility. Made sure the icon shows up on dock. Made fonts work like we originally had it and removed bundling. Reformatted cmake to look more like our style.

* more changes to font manager.

* Reverted changed to sample.

* fix: added .DS_Store to gitignore.

* removed ds_store.

---------

+++ src/controller/src/system/rocprofvis_controller_ext_data.cpp
@@ -211,7 +211,7 @@ rocprofvis_result_t ExtData::GetString(rocprofvis_property_t property, uint64_t 
 ArgumentData::ArgumentData(const char* category, const char* name, const char* value,
     rocprofvis_db_data_type_t type, uint64_t category_enum, size_t position, const char* arg_type) 
     :ExtData(category,name,value, type, category_enum)
-    ,m_position(position)
+    ,m_position(static_cast<uint64_t>(position))
     ,m_arg_type(arg_type) { }
 
 ArgumentData::~ArgumentData() {}
